+++ src-tauri/.gitignore
@@ -2,6 +2,10 @@
 # will have compiled files and executables
 /target/
 
+# Generated by Tauri
+# will have schema files for capabilities auto-completion
+/gen/schemas
+
 # Local environment files (secrets)
 .env
 .env.*
