+++ src-tauri/gen/android/app/src/main/java/com/pawnappetit/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.pawnappetit
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+
+class MainActivity : TauriActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    enableEdgeToEdge()
+    super.onCreate(savedInstanceState)
+  }
+}
