+++ src-tauri/gen/android/app/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">Pawn Appetit</string>
+    <string name="main_activity_title">Pawn Appetit</string>
+</resources>
