+++ src/bindings/generated.ts
@@ -5,6 +5,9 @@
 
 
 export const commands = {
+/**
+ * Platform-specific splashscreen handling
+ */
 async closeSplashscreen() : Promise<Result<null, string>> {
     try {
     return { status: "ok", data: await TAURI_INVOKE("close_splashscreen") };
