+++ vite.config.ts
@@ -7,11 +7,26 @@ import { defineConfig } from "vite";
 
 const isDebug = !!process.env.TAURI_ENV_DEBUG;
 
+const host = process.env.TAURI_DEV_HOST;
+
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react(), vanillaExtractPlugin(), TanStackRouterVite()],
+  clearScreen: false,
   server: {
     port: 1420,
+    strictPort: true,
+    host: host || false,
+    hmr: host
+      ? {
+          protocol: "ws",
+          host,
+          port: 1421,
+        }
+      : undefined,
+    watch: {
+      ignored: ["**/src-tauri/**"],
+    },
   },
   build: {
     minify: isDebug ? false : "esbuild",
